+++ tasmota/settings.h
@@ -343,39 +343,16 @@ struct {
 
   // Start of char array storing all parameter strings ********
 
-  char          text_pool[101];            // 017  Was ota_url[101] - size is settings_text_size
-
-  char          ex_mqtt_prefix[3][11];     // 07C  Free since 8.0.0.1
-  uint8_t       ex_baudrate;               // 09D
-  uint8_t       ex_seriallog_level;        // 09E
-  uint8_t       ex_sta_config;             // 09F
-  uint8_t       ex_sta_active;             // 0A0
+  char          text_pool[138];            // 017  Size is settings_text_size
   char          ex_sta_ssid[2][33];        // 0A1
   char          ex_sta_pwd[2][65];         // 0E3
-  char          ex_hostname[33];           // 165
-  char          ex_syslog_host[33];        // 186
-  uint8_t       ex_rule_stop;              // 1A7
-  uint16_t      ex_syslog_port;            // 1A8
-  uint8_t       ex_syslog_level;           // 1AA
-  uint8_t       ex_webserver;              // 1AB
-  uint8_t       ex_weblog_level;           // 1AC
-  uint8_t       ex_mqtt_fingerprint[2][20];  // 1AD
-  uint8_t       ex_ex_adc_param_type;      // 1D5
-  uint8_t       ex_free_1d6[10];           // 1D6
-  SysBitfield4  ex_flag4;                  // 1E0
-  uint8_t       ex_serial_config;          // 1E4
-  uint8_t       ex_wifi_output_power;      // 1E5
-  uint8_t       ex_shutter_accuracy;       // 1E6
-  uint8_t       ex_mqttlog_level;          // 1E7
-  uint8_t       ex_sps30_inuse_hours;      // 1E8
+  uint8_t       text_pool_165[132];        // 165
   char          ex_mqtt_host[33];          // 1E9
-  uint16_t      ex_mqtt_port;              // 20A
-  char          ex_mqtt_client[33];        // 20C
+  uint8_t       text_pool_20A[35];         // 20A
   char          ex_mqtt_user[33];          // 22D
   char          ex_mqtt_pwd[33];           // 24E
   char          ex_mqtt_topic[33];         // 26F
-  char          ex_button_topic[33];       // 290
-  char          ex_mqtt_grptopic[33];      // 2B1
+  uint8_t       text_pool_290[66];         // 290
 
   // End of single char array of 698 chars max ****************
 
@@ -399,7 +376,7 @@ struct {
   int16_t       toffset[2];                // 30E
   uint8_t       display_font;              // 312
 
-  char          ex_state_text[4][11];      // 313  Free since 8.0.0.1
+  uint8_t       free_313[44];              // 313
 
   uint8_t       tuyamcu_topic;             // 33F  Manage tuyaSend topic. ex_energy_power_delta on 6.6.0.20, replaced on 8.5.0.1
   uint16_t      domoticz_update_timer;     // 340
@@ -433,7 +410,9 @@ struct {
   uint16_t      blinkcount;                // 39C
   uint16_t      light_rotation;            // 39E
   SysBitfield3  flag3;                     // 3A0
-  uint8_t       switchmode[MAX_SWITCHES];  // 3A4
+
+  uint8_t       ex_switchmode[8];          // 3A4 - Free since 9.2.0.6
+
   myio          my_gp;                     // 3AC  2 x 18 bytes (ESP8266) / 2 x 40 bytes (ESP32)
 #ifdef ESP8266
   uint16_t      gpio16_converted;          // 3D0
@@ -487,12 +466,17 @@ struct {
   uint8_t       knx_GA_registered;         // 4A5  Number of Group Address to read
   uint16_t      light_wakeup;              // 4A6
   uint8_t       knx_CB_registered;         // 4A8  Number of Group Address to write
+  uint8_t       switchmode[MAX_SWITCHES_SET];  // 4A9
 
-  char          ex_web_password[33];       // 4A9  Free since 8.0.0.1
+  uint8_t       free_4c5[5];               // 4C5
 
-  uint8_t       interlock[MAX_INTERLOCKS]; // 4CA
+  uint8_t       ex_interlock[4];           // 4CA MAX_INTERLOCKS = MAX_RELAYS / 2 (Legacy)
 
-  char          ex_ntp_server[3][33];      // 4CE  Free since 8.0.0.1
+  uint8_t       free_4ce[2];               // 4CE
+
+  power_t       interlock[MAX_INTERLOCKS_SET];  // 4D0 MAX_INTERLOCKS = MAX_RELAYS / 2
+
+  uint8_t       free_508[41];              // 508
 
   uint8_t       ina219_mode;               // 531
   uint16_t      pulse_timer[MAX_PULSETIMERS];  // 532
@@ -500,7 +484,7 @@ struct {
   uint32_t      ipv4_address[4];           // 544
   unsigned long energy_kWhtotal;           // 554
 
-  char          ex_mqtt_fulltopic[100];    // 558  Free since 8.0.0.1
+  uint8_t       free_558[100];             // 558
 
   SysBitfield2  flag2;                     // 5BC
   unsigned long pulse_counter[MAX_COUNTERS];  // 5C0
@@ -586,7 +570,7 @@ struct {
   uint8_t       weight_change;             // E9F
   uint8_t       web_color2[2][3];          // EA0  Needs to be on integer / 3 distance from web_color
 
-  char          ex_cors_domain[33];        // EA6  Free since 8.0.0.1
+  uint8_t       free_ea6[33];              // EA6
 
   uint8_t       sta_config;                // EC7
   uint8_t       sta_active;                // EC8
@@ -661,7 +645,7 @@ struct {
 
   int           ex_adc_param4;             // FD8  Free since 9.0.0.1
 
-  uint32_t      shutter_button[MAX_KEYS];  // FDC
+  uint32_t      shutter_button[MAX_SHUTTER_KEYS];  // FDC
   uint32_t      i2c_drivers[3];            // FEC  I2cDriver
   uint32_t      cfg_timestamp;             // FF8
   uint32_t      cfg_crc32;                 // FFC